requirements.txt
asgiref             3.7.2
Django              4.2.5
djangorestframework 3.14.0
pip                 23.2.1
psycopg2            2.9.7
pytz                2023.3.post1
setuptools          65.5.0
